removed kafkaLiteComponent from assembly alias
