+++ engine/base/kafka-runtime-bin-test/src/it/scala/pl/touk/nussknacker/engine/baseengine/kafka/NuKafkaEngineBinTest.scala
