--- engine/base/nu-kafka-engine-bin-test/src/it/resources/scenario.json
@@ -1,133 +1,0 @@
-{
-  "metaData" : {
-    "id" : "DetectLargeTransactions",
-    "typeSpecificData" : {
-      "parallelism" : 2,
-      "spillStateToDisk" : true,
-      "useAsyncInterpretation" : null,
-      "checkpointIntervalInSeconds" : null,
-      "type" : "StreamMetaData"
-    },
-    "isSubprocess" : false,
-    "additionalFields" : {
-      "description" : null,
-      "groups" : [
-      ],
-      "properties" : {
-
-      }
-    },
-    "subprocessVersions" : {
-
-    }
-  },
-  "exceptionHandlerRef" : {
-    "parameters" : [
-    ]
-  },
-  "nodes" : [
-    {
-      "id" : "transactions",
-      "ref" : {
-        "typ" : "kafka-registry-typed-json",
-        "parameters" : [
-          {
-            "name" : "Topic",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "'transactions'"
-            }
-          },
-          {
-            "name" : "Schema version",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "'latest'"
-            }
-          }
-        ]
-      },
-      "additionalFields" : {
-        "description" : null,
-        "layoutData" : {
-          "x" : 0,
-          "y" : 0
-        }
-      },
-      "type" : "Source"
-    },
-    {
-      "nextFalse" : [
-      ],
-      "id" : "only large ones",
-      "expression" : {
-        "language" : "spel",
-        "expression" : "#input.amount > 20"
-      },
-      "isDisabled" : null,
-      "additionalFields" : {
-        "description" : null,
-        "layoutData" : {
-          "x" : 0,
-          "y" : 180
-        }
-      },
-      "type" : "Filter"
-    },
-    {
-      "id" : "send for audit",
-      "ref" : {
-        "typ" : "kafka-registry-typed-json-raw",
-        "parameters" : [
-          {
-            "name" : "Topic",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "'processedEvents'"
-            }
-          },
-          {
-            "name" : "Schema version",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "'latest'"
-            }
-          },
-          {
-            "name" : "Value validation mode",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "'strict'"
-            }
-          },
-          {
-            "name" : "Key",
-            "expression" : {
-              "language" : "spel",
-              "expression" : ""
-            }
-          },
-          {
-            "name" : "Value",
-            "expression" : {
-              "language" : "spel",
-              "expression" : "#input"
-            }
-          }
-        ]
-      },
-      "endResult" : null,
-      "isDisabled" : null,
-      "additionalFields" : {
-        "description" : null,
-        "layoutData" : {
-          "x" : 0,
-          "y" : 360
-        }
-      },
-      "type" : "Sink"
-    }
-  ],
-  "additionalBranches" : [
-  ]
-}
